requirements.txt
# =============================================================================
# requirements.txt — Dependências do sistema de multiagentes
# =============================================================================
# Instale com: pip install -r requirements.txt

# ---------------------------------------------------------------------------
# OpenAI SDK (client oficial — GPT-4o-mini)
# ---------------------------------------------------------------------------
openai>=1.51.0

# ---------------------------------------------------------------------------
# Variáveis de ambiente (.env)
# ---------------------------------------------------------------------------
python-dotenv>=1.0.0

# ---------------------------------------------------------------------------
# Requisições HTTP (Notion REST API)
# ---------------------------------------------------------------------------
requests>=2.31.0

# ---------------------------------------------------------------------------
# Agendamento do loop do Focus Guard
# ---------------------------------------------------------------------------
schedule>=1.2.1

# ---------------------------------------------------------------------------
# Utilitários de data/hora (já incluído no stdlib, mas explicitar versão)
# ---------------------------------------------------------------------------
# python-dateutil>=2.8.2   # descomente se precisar de parsing avançado de datas

# ---------------------------------------------------------------------------
# Colorama (compatibilidade de cores ANSI no Windows)
# ---------------------------------------------------------------------------
colorama>=0.4.6

# ---------------------------------------------------------------------------
# Retry com backoff exponencial (rate limit da Notion API)
# ---------------------------------------------------------------------------
tenacity>=8.2.0

# ---------------------------------------------------------------------------
# Testes automatizados
# ---------------------------------------------------------------------------
pytest>=8.0.0
pytest-mock>=3.14.0
fakeredis>=2.23.0

# ---------------------------------------------------------------------------
# Opcional: rich (terminal mais bonito) — descomente para usar
# ---------------------------------------------------------------------------
# rich>=13.7.0

# ---------------------------------------------------------------------------
# Web interface (FastAPI + HTMX)
# ---------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# ---------------------------------------------------------------------------
# Google Calendar
# ---------------------------------------------------------------------------
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
tzlocal>=5.0

# ---------------------------------------------------------------------------
# Retrospective page support
# ---------------------------------------------------------------------------
anyio>=4.0.0

# ---------------------------------------------------------------------------
# Redis (persistência durável — Railway)
# ---------------------------------------------------------------------------
redis>=5.0.0
